+++ examples/complete/main.tf
@@ -64,6 +64,10 @@ module "vpc" {
       bgp_asn    = 65112
       ip_address = "5.6.7.8"
     }
+    IP3 = {
+      bgp_asn_extended = 2147483648
+      ip_address       = "5.6.7.8"
+    }
   }
 
   enable_vpn_gateway = true
@@ -71,12 +75,6 @@ module "vpc" {
   enable_dhcp_options              = true
   dhcp_options_domain_name         = "service.consul"
   dhcp_options_domain_name_servers = ["127.0.0.1", "10.10.0.2"]
-
-  # VPC Flow Logs (Cloudwatch log group and IAM role will be created)
-  enable_flow_log                      = true
-  create_flow_log_cloudwatch_log_group = true
-  create_flow_log_cloudwatch_iam_role  = true
-  flow_log_max_aggregation_interval    = 60
 
   tags = local.tags
 }
@@ -102,8 +100,12 @@ module "vpc_endpoints" {
 
   endpoints = {
     s3 = {
-      service = "s3"
-      tags    = { Name = "s3-vpc-endpoint" }
+      service             = "s3"
+      private_dns_enabled = true
+      dns_options = {
+        private_dns_only_for_inbound_resolver_endpoint = false
+      }
+      tags = { Name = "s3-vpc-endpoint" }
     },
     dynamodb = {
       service         = "dynamodb"
@@ -116,6 +118,13 @@ module "vpc_endpoints" {
       service             = "ecs"
       private_dns_enabled = true
       subnet_ids          = module.vpc.private_subnets
+      subnet_configurations = [
+        for v in module.vpc.private_subnet_objects :
+        {
+          ipv4      = cidrhost(v.cidr_block, 10)
+          subnet_id = v.id
+        }
+      ]
     },
     ecs_telemetry = {
       create              = false