+++ main.tf
@@ -28,6 +28,8 @@ locals {
 resource "aws_vpc" "this" {
   count = local.create_vpc ? 1 : 0
 
+  region = var.region
+
   cidr_block          = var.use_ipam_pool ? null : var.cidr
   ipv4_ipam_pool_id   = var.ipv4_ipam_pool_id
   ipv4_netmask_length = var.ipv4_netmask_length
@@ -53,10 +55,49 @@ resource "aws_vpc" "this" {
 resource "aws_vpc_ipv4_cidr_block_association" "this" {
   count = local.create_vpc && length(var.secondary_cidr_blocks) > 0 ? length(var.secondary_cidr_blocks) : 0
 
+  region = var.region
+
   # Do not turn this into `local.vpc_id`
   vpc_id = aws_vpc.this[0].id
 
   cidr_block = element(var.secondary_cidr_blocks, count.index)
+}
+
+resource "aws_vpc_block_public_access_options" "this" {
+  count = local.create_vpc && length(keys(var.vpc_block_public_access_options)) > 0 ? 1 : 0
+
+  region = var.region
+
+  internet_gateway_block_mode = try(var.vpc_block_public_access_options["internet_gateway_block_mode"], null)
+}
+
+resource "aws_vpc_block_public_access_exclusion" "this" {
+  for_each = { for k, v in var.vpc_block_public_access_exclusions : k => v if local.create_vpc }
+
+  region = var.region
+
+  vpc_id = try(each.value.exclude_vpc, false) ? local.vpc_id : null
+
+  subnet_id = try(each.value.exclude_subnet, false) ? lookup(
+    {
+      private     = aws_subnet.private[*].id,
+      public      = aws_subnet.public[*].id,
+      database    = aws_subnet.database[*].id,
+      redshift    = aws_subnet.redshift[*].id,
+      elasticache = aws_subnet.elasticache[*].id,
+      intra       = aws_subnet.intra[*].id,
+      outpost     = aws_subnet.outpost[*].id
+    },
+    each.value.subnet_type,
+    null
+  )[each.value.subnet_index] : null
+
+  internet_gateway_exclusion_mode = each.value.internet_gateway_exclusion_mode
+
+  tags = merge(
+    var.tags,
+    try(each.value.tags, {}),
+  )
 }
 
 ################################################################################
@@ -66,11 +107,14 @@ resource "aws_vpc_ipv4_cidr_block_association" "this" {
 resource "aws_vpc_dhcp_options" "this" {
   count = local.create_vpc && var.enable_dhcp_options ? 1 : 0
 
-  domain_name          = var.dhcp_options_domain_name
-  domain_name_servers  = var.dhcp_options_domain_name_servers
-  ntp_servers          = var.dhcp_options_ntp_servers
-  netbios_name_servers = var.dhcp_options_netbios_name_servers
-  netbios_node_type    = var.dhcp_options_netbios_node_type
+  region = var.region
+
+  domain_name                       = var.dhcp_options_domain_name
+  domain_name_servers               = var.dhcp_options_domain_name_servers
+  ntp_servers                       = var.dhcp_options_ntp_servers
+  netbios_name_servers              = var.dhcp_options_netbios_name_servers
+  netbios_node_type                 = var.dhcp_options_netbios_node_type
+  ipv6_address_preferred_lease_time = var.dhcp_options_ipv6_address_preferred_lease_time
 
   tags = merge(
     { "Name" = var.name },
@@ -82,20 +126,26 @@ resource "aws_vpc_dhcp_options" "this" {
 resource "aws_vpc_dhcp_options_association" "this" {
   count = local.create_vpc && var.enable_dhcp_options ? 1 : 0
 
+  region = var.region
+
   vpc_id          = local.vpc_id
   dhcp_options_id = aws_vpc_dhcp_options.this[0].id
 }
 
 ################################################################################
-# Publiс Subnets
+# Public Subnets
 ################################################################################
 
 locals {
   create_public_subnets = local.create_vpc && local.len_public_subnets > 0
+
+  num_public_route_tables = var.create_multiple_public_route_tables ? local.len_public_subnets : 1
 }
 
 resource "aws_subnet" "public" {
   count = local.create_public_subnets && (!var.one_nat_gateway_per_az || local.len_public_subnets >= length(var.azs)) ? local.len_public_subnets : 0
+
+  region = var.region
 
   assign_ipv6_address_on_creation                = var.enable_ipv6 && var.public_subnet_ipv6_native ? true : var.public_subnet_assign_ipv6_address_on_creation
   availability_zone                              = length(regexall("^[a-z]{2}-", element(var.azs, count.index))) > 0 ? element(var.azs, count.index) : null
@@ -124,12 +174,19 @@ resource "aws_subnet" "public" {
 }
 
 resource "aws_route_table" "public" {
-  count = local.create_public_subnets ? 1 : 0
+  count = local.create_public_subnets ? local.num_public_route_tables : 0
+
+  region = var.region
 
   vpc_id = local.vpc_id
 
   tags = merge(
-    { "Name" = "${var.name}-${var.public_subnet_suffix}" },
+    {
+      "Name" = var.create_multiple_public_route_tables ? format(
+        "${var.name}-${var.public_subnet_suffix}-%s",
+        element(var.azs, count.index),
+      ) : "${var.name}-${var.public_subnet_suffix}"
+    },
     var.tags,
     var.public_route_table_tags,
   )
@@ -138,14 +195,18 @@ resource "aws_route_table" "public" {
 resource "aws_route_table_association" "public" {
   count = local.create_public_subnets ? local.len_public_subnets : 0
 
+  region = var.region
+
   subnet_id      = element(aws_subnet.public[*].id, count.index)
-  route_table_id = aws_route_table.public[0].id
+  route_table_id = element(aws_route_table.public[*].id, var.create_multiple_public_route_tables ? count.index : 0)
 }
 
 resource "aws_route" "public_internet_gateway" {
-  count = local.create_public_subnets && var.create_igw ? 1 : 0
+  count = local.create_public_subnets && var.create_igw ? local.num_public_route_tables : 0
 
-  route_table_id         = aws_route_table.public[0].id
+  region = var.region
+
+  route_table_id         = aws_route_table.public[count.index].id
   destination_cidr_block = "0.0.0.0/0"
   gateway_id             = aws_internet_gateway.this[0].id
 
@@ -155,9 +216,11 @@ resource "aws_route" "public_internet_gateway" {
 }
 
 resource "aws_route" "public_internet_gateway_ipv6" {
-  count = local.create_public_subnets && var.create_igw && var.enable_ipv6 ? 1 : 0
+  count = local.create_public_subnets && var.create_igw && var.enable_ipv6 ? local.num_public_route_tables : 0
 
-  route_table_id              = aws_route_table.public[0].id
+  region = var.region
+
+  route_table_id              = aws_route_table.public[count.index].id
   destination_ipv6_cidr_block = "::/0"
   gateway_id                  = aws_internet_gateway.this[0].id
 }
@@ -168,6 +231,8 @@ resource "aws_route" "public_internet_gateway_ipv6" {
 
 resource "aws_network_acl" "public" {
   count = local.create_public_subnets && var.public_dedicated_network_acl ? 1 : 0
+
+  region = var.region
 
   vpc_id     = local.vpc_id
   subnet_ids = aws_subnet.public[*].id
@@ -181,6 +246,8 @@ resource "aws_network_acl" "public" {
 
 resource "aws_network_acl_rule" "public_inbound" {
   count = local.create_public_subnets && var.public_dedicated_network_acl ? length(var.public_inbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.public[0].id
 
@@ -198,6 +265,8 @@ resource "aws_network_acl_rule" "public_inbound" {
 
 resource "aws_network_acl_rule" "public_outbound" {
   count = local.create_public_subnets && var.public_dedicated_network_acl ? length(var.public_outbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.public[0].id
 
@@ -223,6 +292,8 @@ locals {
 
 resource "aws_subnet" "private" {
   count = local.create_private_subnets ? local.len_private_subnets : 0
+
+  region = var.region
 
   assign_ipv6_address_on_creation                = var.enable_ipv6 && var.private_subnet_ipv6_native ? true : var.private_subnet_assign_ipv6_address_on_creation
   availability_zone                              = length(regexall("^[a-z]{2}-", element(var.azs, count.index))) > 0 ? element(var.azs, count.index) : null
@@ -253,6 +324,8 @@ resource "aws_subnet" "private" {
 resource "aws_route_table" "private" {
   count = local.create_private_subnets && local.max_subnet_length > 0 ? local.nat_gateway_count : 0
 
+  region = var.region
+
   vpc_id = local.vpc_id
 
   tags = merge(
@@ -269,6 +342,8 @@ resource "aws_route_table" "private" {
 
 resource "aws_route_table_association" "private" {
   count = local.create_private_subnets ? local.len_private_subnets : 0
+
+  region = var.region
 
   subnet_id = element(aws_subnet.private[*].id, count.index)
   route_table_id = element(
@@ -288,6 +363,8 @@ locals {
 resource "aws_network_acl" "private" {
   count = local.create_private_network_acl ? 1 : 0
 
+  region = var.region
+
   vpc_id     = local.vpc_id
   subnet_ids = aws_subnet.private[*].id
 
@@ -300,6 +377,8 @@ resource "aws_network_acl" "private" {
 
 resource "aws_network_acl_rule" "private_inbound" {
   count = local.create_private_network_acl ? length(var.private_inbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.private[0].id
 
@@ -317,6 +396,8 @@ resource "aws_network_acl_rule" "private_inbound" {
 
 resource "aws_network_acl_rule" "private_outbound" {
   count = local.create_private_network_acl ? length(var.private_outbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.private[0].id
 
@@ -343,6 +424,8 @@ locals {
 
 resource "aws_subnet" "database" {
   count = local.create_database_subnets ? local.len_database_subnets : 0
+
+  region = var.region
 
   assign_ipv6_address_on_creation                = var.enable_ipv6 && var.database_subnet_ipv6_native ? true : var.database_subnet_assign_ipv6_address_on_creation
   availability_zone                              = length(regexall("^[a-z]{2}-", element(var.azs, count.index))) > 0 ? element(var.azs, count.index) : null
@@ -371,6 +454,8 @@ resource "aws_subnet" "database" {
 resource "aws_db_subnet_group" "database" {
   count = local.create_database_subnets && var.create_database_subnet_group ? 1 : 0
 
+  region = var.region
+
   name        = lower(coalesce(var.database_subnet_group_name, var.name))
   description = "Database subnet group for ${var.name}"
   subnet_ids  = aws_subnet.database[*].id
@@ -386,6 +471,8 @@ resource "aws_db_subnet_group" "database" {
 
 resource "aws_route_table" "database" {
   count = local.create_database_route_table ? var.single_nat_gateway || var.create_database_internet_gateway_route ? 1 : local.len_database_subnets : 0
+
+  region = var.region
 
   vpc_id = local.vpc_id
 
@@ -404,6 +491,8 @@ resource "aws_route_table" "database" {
 resource "aws_route_table_association" "database" {
   count = local.create_database_subnets ? local.len_database_subnets : 0
 
+  region = var.region
+
   subnet_id = element(aws_subnet.database[*].id, count.index)
   route_table_id = element(
     coalescelist(aws_route_table.database[*].id, aws_route_table.private[*].id),
@@ -413,6 +502,8 @@ resource "aws_route_table_association" "database" {
 
 resource "aws_route" "database_internet_gateway" {
   count = local.create_database_route_table && var.create_igw && var.create_database_internet_gateway_route && !var.create_database_nat_gateway_route ? 1 : 0
+
+  region = var.region
 
   route_table_id         = aws_route_table.database[0].id
   destination_cidr_block = "0.0.0.0/0"
@@ -426,6 +517,8 @@ resource "aws_route" "database_internet_gateway" {
 resource "aws_route" "database_nat_gateway" {
   count = local.create_database_route_table && !var.create_database_internet_gateway_route && var.create_database_nat_gateway_route && var.enable_nat_gateway ? var.single_nat_gateway ? 1 : local.len_database_subnets : 0
 
+  region = var.region
+
   route_table_id         = element(aws_route_table.database[*].id, count.index)
   destination_cidr_block = "0.0.0.0/0"
   nat_gateway_id         = element(aws_nat_gateway.this[*].id, count.index)
@@ -438,6 +531,8 @@ resource "aws_route" "database_nat_gateway" {
 resource "aws_route" "database_dns64_nat_gateway" {
   count = local.create_database_route_table && !var.create_database_internet_gateway_route && var.create_database_nat_gateway_route && var.enable_nat_gateway && var.enable_ipv6 && var.private_subnet_enable_dns64 ? var.single_nat_gateway ? 1 : local.len_database_subnets : 0
 
+  region = var.region
+
   route_table_id              = element(aws_route_table.database[*].id, count.index)
   destination_ipv6_cidr_block = "64:ff9b::/96"
   nat_gateway_id              = element(aws_nat_gateway.this[*].id, count.index)
@@ -449,6 +544,8 @@ resource "aws_route" "database_dns64_nat_gateway" {
 
 resource "aws_route" "database_ipv6_egress" {
   count = local.create_database_route_table && var.create_egress_only_igw && var.enable_ipv6 && var.create_database_internet_gateway_route ? 1 : 0
+
+  region = var.region
 
   route_table_id              = aws_route_table.database[0].id
   destination_ipv6_cidr_block = "::/0"
@@ -470,6 +567,8 @@ locals {
 resource "aws_network_acl" "database" {
   count = local.create_database_network_acl ? 1 : 0
 
+  region = var.region
+
   vpc_id     = local.vpc_id
   subnet_ids = aws_subnet.database[*].id
 
@@ -482,6 +581,8 @@ resource "aws_network_acl" "database" {
 
 resource "aws_network_acl_rule" "database_inbound" {
   count = local.create_database_network_acl ? length(var.database_inbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.database[0].id
 
@@ -499,6 +600,8 @@ resource "aws_network_acl_rule" "database_inbound" {
 
 resource "aws_network_acl_rule" "database_outbound" {
   count = local.create_database_network_acl ? length(var.database_outbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.database[0].id
 
@@ -525,6 +628,8 @@ locals {
 
 resource "aws_subnet" "redshift" {
   count = local.create_redshift_subnets ? local.len_redshift_subnets : 0
+
+  region = var.region
 
   assign_ipv6_address_on_creation                = var.enable_ipv6 && var.redshift_subnet_ipv6_native ? true : var.redshift_subnet_assign_ipv6_address_on_creation
   availability_zone                              = length(regexall("^[a-z]{2}-", element(var.azs, count.index))) > 0 ? element(var.azs, count.index) : null
@@ -553,6 +658,8 @@ resource "aws_subnet" "redshift" {
 resource "aws_redshift_subnet_group" "redshift" {
   count = local.create_redshift_subnets && var.create_redshift_subnet_group ? 1 : 0
 
+  region = var.region
+
   name        = lower(coalesce(var.redshift_subnet_group_name, var.name))
   description = "Redshift subnet group for ${var.name}"
   subnet_ids  = aws_subnet.redshift[*].id
@@ -567,6 +674,8 @@ resource "aws_redshift_subnet_group" "redshift" {
 resource "aws_route_table" "redshift" {
   count = local.create_redshift_route_table ? 1 : 0
 
+  region = var.region
+
   vpc_id = local.vpc_id
 
   tags = merge(
@@ -579,6 +688,8 @@ resource "aws_route_table" "redshift" {
 resource "aws_route_table_association" "redshift" {
   count = local.create_redshift_subnets && !var.enable_public_redshift ? local.len_redshift_subnets : 0
 
+  region = var.region
+
   subnet_id = element(aws_subnet.redshift[*].id, count.index)
   route_table_id = element(
     coalescelist(aws_route_table.redshift[*].id, aws_route_table.private[*].id),
@@ -588,6 +699,8 @@ resource "aws_route_table_association" "redshift" {
 
 resource "aws_route_table_association" "redshift_public" {
   count = local.create_redshift_subnets && var.enable_public_redshift ? local.len_redshift_subnets : 0
+
+  region = var.region
 
   subnet_id = element(aws_subnet.redshift[*].id, count.index)
   route_table_id = element(
@@ -607,6 +720,8 @@ locals {
 resource "aws_network_acl" "redshift" {
   count = local.create_redshift_network_acl ? 1 : 0
 
+  region = var.region
+
   vpc_id     = local.vpc_id
   subnet_ids = aws_subnet.redshift[*].id
 
@@ -619,6 +734,8 @@ resource "aws_network_acl" "redshift" {
 
 resource "aws_network_acl_rule" "redshift_inbound" {
   count = local.create_redshift_network_acl ? length(var.redshift_inbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.redshift[0].id
 
@@ -636,6 +753,8 @@ resource "aws_network_acl_rule" "redshift_inbound" {
 
 resource "aws_network_acl_rule" "redshift_outbound" {
   count = local.create_redshift_network_acl ? length(var.redshift_outbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.redshift[0].id
 
@@ -662,6 +781,8 @@ locals {
 
 resource "aws_subnet" "elasticache" {
   count = local.create_elasticache_subnets ? local.len_elasticache_subnets : 0
+
+  region = var.region
 
   assign_ipv6_address_on_creation                = var.enable_ipv6 && var.elasticache_subnet_ipv6_native ? true : var.elasticache_subnet_assign_ipv6_address_on_creation
   availability_zone                              = length(regexall("^[a-z]{2}-", element(var.azs, count.index))) > 0 ? element(var.azs, count.index) : null
@@ -690,6 +811,8 @@ resource "aws_subnet" "elasticache" {
 resource "aws_elasticache_subnet_group" "elasticache" {
   count = local.create_elasticache_subnets && var.create_elasticache_subnet_group ? 1 : 0
 
+  region = var.region
+
   name        = coalesce(var.elasticache_subnet_group_name, var.name)
   description = "ElastiCache subnet group for ${var.name}"
   subnet_ids  = aws_subnet.elasticache[*].id
@@ -704,6 +827,8 @@ resource "aws_elasticache_subnet_group" "elasticache" {
 resource "aws_route_table" "elasticache" {
   count = local.create_elasticache_route_table ? 1 : 0
 
+  region = var.region
+
   vpc_id = local.vpc_id
 
   tags = merge(
@@ -715,6 +840,8 @@ resource "aws_route_table" "elasticache" {
 
 resource "aws_route_table_association" "elasticache" {
   count = local.create_elasticache_subnets ? local.len_elasticache_subnets : 0
+
+  region = var.region
 
   subnet_id = element(aws_subnet.elasticache[*].id, count.index)
   route_table_id = element(
@@ -737,6 +864,8 @@ locals {
 resource "aws_network_acl" "elasticache" {
   count = local.create_elasticache_network_acl ? 1 : 0
 
+  region = var.region
+
   vpc_id     = local.vpc_id
   subnet_ids = aws_subnet.elasticache[*].id
 
@@ -749,6 +878,8 @@ resource "aws_network_acl" "elasticache" {
 
 resource "aws_network_acl_rule" "elasticache_inbound" {
   count = local.create_elasticache_network_acl ? length(var.elasticache_inbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.elasticache[0].id
 
@@ -766,6 +897,8 @@ resource "aws_network_acl_rule" "elasticache_inbound" {
 
 resource "aws_network_acl_rule" "elasticache_outbound" {
   count = local.create_elasticache_network_acl ? length(var.elasticache_outbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.elasticache[0].id
 
@@ -786,11 +919,14 @@ resource "aws_network_acl_rule" "elasticache_outbound" {
 ################################################################################
 
 locals {
-  create_intra_subnets = local.create_vpc && local.len_intra_subnets > 0
+  create_intra_subnets   = local.create_vpc && local.len_intra_subnets > 0
+  num_intra_route_tables = var.create_multiple_intra_route_tables ? local.len_intra_subnets : 1
 }
 
 resource "aws_subnet" "intra" {
   count = local.create_intra_subnets ? local.len_intra_subnets : 0
+
+  region = var.region
 
   assign_ipv6_address_on_creation                = var.enable_ipv6 && var.intra_subnet_ipv6_native ? true : var.intra_subnet_assign_ipv6_address_on_creation
   availability_zone                              = length(regexall("^[a-z]{2}-", element(var.azs, count.index))) > 0 ? element(var.azs, count.index) : null
@@ -817,12 +953,19 @@ resource "aws_subnet" "intra" {
 }
 
 resource "aws_route_table" "intra" {
-  count = local.create_intra_subnets ? 1 : 0
+  count = local.create_intra_subnets ? local.num_intra_route_tables : 0
+
+  region = var.region
 
   vpc_id = local.vpc_id
 
   tags = merge(
-    { "Name" = "${var.name}-${var.intra_subnet_suffix}" },
+    {
+      "Name" = var.create_multiple_intra_route_tables ? format(
+        "${var.name}-${var.intra_subnet_suffix}-%s",
+        element(var.azs, count.index),
+      ) : "${var.name}-${var.intra_subnet_suffix}"
+    },
     var.tags,
     var.intra_route_table_tags,
   )
@@ -831,8 +974,10 @@ resource "aws_route_table" "intra" {
 resource "aws_route_table_association" "intra" {
   count = local.create_intra_subnets ? local.len_intra_subnets : 0
 
+  region = var.region
+
   subnet_id      = element(aws_subnet.intra[*].id, count.index)
-  route_table_id = element(aws_route_table.intra[*].id, 0)
+  route_table_id = element(aws_route_table.intra[*].id, var.create_multiple_intra_route_tables ? count.index : 0)
 }
 
 ################################################################################
@@ -846,6 +991,8 @@ locals {
 resource "aws_network_acl" "intra" {
   count = local.create_intra_network_acl ? 1 : 0
 
+  region = var.region
+
   vpc_id     = local.vpc_id
   subnet_ids = aws_subnet.intra[*].id
 
@@ -858,6 +1005,8 @@ resource "aws_network_acl" "intra" {
 
 resource "aws_network_acl_rule" "intra_inbound" {
   count = local.create_intra_network_acl ? length(var.intra_inbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.intra[0].id
 
@@ -875,6 +1024,8 @@ resource "aws_network_acl_rule" "intra_inbound" {
 
 resource "aws_network_acl_rule" "intra_outbound" {
   count = local.create_intra_network_acl ? length(var.intra_outbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.intra[0].id
 
@@ -900,6 +1051,8 @@ locals {
 
 resource "aws_subnet" "outpost" {
   count = local.create_outpost_subnets ? local.len_outpost_subnets : 0
+
+  region = var.region
 
   assign_ipv6_address_on_creation                = var.enable_ipv6 && var.outpost_subnet_ipv6_native ? true : var.outpost_subnet_assign_ipv6_address_on_creation
   availability_zone                              = var.outpost_az
@@ -930,6 +1083,8 @@ resource "aws_subnet" "outpost" {
 resource "aws_route_table_association" "outpost" {
   count = local.create_outpost_subnets ? local.len_outpost_subnets : 0
 
+  region = var.region
+
   subnet_id = element(aws_subnet.outpost[*].id, count.index)
   route_table_id = element(
     aws_route_table.private[*].id,
@@ -948,6 +1103,8 @@ locals {
 resource "aws_network_acl" "outpost" {
   count = local.create_outpost_network_acl ? 1 : 0
 
+  region = var.region
+
   vpc_id     = local.vpc_id
   subnet_ids = aws_subnet.outpost[*].id
 
@@ -960,6 +1117,8 @@ resource "aws_network_acl" "outpost" {
 
 resource "aws_network_acl_rule" "outpost_inbound" {
   count = local.create_outpost_network_acl ? length(var.outpost_inbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.outpost[0].id
 
@@ -977,6 +1136,8 @@ resource "aws_network_acl_rule" "outpost_inbound" {
 
 resource "aws_network_acl_rule" "outpost_outbound" {
   count = local.create_outpost_network_acl ? length(var.outpost_outbound_acl_rules) : 0
+
+  region = var.region
 
   network_acl_id = aws_network_acl.outpost[0].id
 
@@ -999,6 +1160,8 @@ resource "aws_network_acl_rule" "outpost_outbound" {
 resource "aws_internet_gateway" "this" {
   count = local.create_public_subnets && var.create_igw ? 1 : 0
 
+  region = var.region
+
   vpc_id = local.vpc_id
 
   tags = merge(
@@ -1011,6 +1174,8 @@ resource "aws_internet_gateway" "this" {
 resource "aws_egress_only_internet_gateway" "this" {
   count = local.create_vpc && var.create_egress_only_igw && var.enable_ipv6 && local.max_subnet_length > 0 ? 1 : 0
 
+  region = var.region
+
   vpc_id = local.vpc_id
 
   tags = merge(
@@ -1021,7 +1186,9 @@ resource "aws_egress_only_internet_gateway" "this" {
 }
 
 resource "aws_route" "private_ipv6_egress" {
-  count = local.create_vpc && var.create_egress_only_igw && var.enable_ipv6 ? local.len_private_subnets : 0
+  count = local.create_vpc && var.create_egress_only_igw && var.enable_ipv6 && local.len_private_subnets > 0 ? local.nat_gateway_count : 0
+
+  region = var.region
 
   route_table_id              = element(aws_route_table.private[*].id, count.index)
   destination_ipv6_cidr_block = "::/0"
@@ -1034,11 +1201,13 @@ resource "aws_route" "private_ipv6_egress" {
 
 locals {
   nat_gateway_count = var.single_nat_gateway ? 1 : var.one_nat_gateway_per_az ? length(var.azs) : local.max_subnet_length
-  nat_gateway_ips   = var.reuse_nat_ips ? var.external_nat_ip_ids : try(aws_eip.nat[*].id, [])
+  nat_gateway_ips   = var.reuse_nat_ips ? var.external_nat_ip_ids : aws_eip.nat[*].id
 }
 
 resource "aws_eip" "nat" {
   count = local.create_vpc && var.enable_nat_gateway && !var.reuse_nat_ips ? local.nat_gateway_count : 0
+
+  region = var.region
 
   domain = "vpc"
 
@@ -1058,6 +1227,8 @@ resource "aws_eip" "nat" {
 
 resource "aws_nat_gateway" "this" {
   count = local.create_vpc && var.enable_nat_gateway ? local.nat_gateway_count : 0
+
+  region = var.region
 
   allocation_id = element(
     local.nat_gateway_ips,
@@ -1083,7 +1254,9 @@ resource "aws_nat_gateway" "this" {
 }
 
 resource "aws_route" "private_nat_gateway" {
-  count = local.create_vpc && var.enable_nat_gateway ? local.nat_gateway_count : 0
+  count = local.create_vpc && var.enable_nat_gateway && var.create_private_nat_gateway_route ? local.nat_gateway_count : 0
+
+  region = var.region
 
   route_table_id         = element(aws_route_table.private[*].id, count.index)
   destination_cidr_block = var.nat_gateway_destination_cidr_block
@@ -1096,6 +1269,8 @@ resource "aws_route" "private_nat_gateway" {
 
 resource "aws_route" "private_dns64_nat_gateway" {
   count = local.create_vpc && var.enable_nat_gateway && var.enable_ipv6 && var.private_subnet_enable_dns64 ? local.nat_gateway_count : 0
+
+  region = var.region
 
   route_table_id              = element(aws_route_table.private[*].id, count.index)
   destination_ipv6_cidr_block = "64:ff9b::/96"
@@ -1113,16 +1288,23 @@ resource "aws_route" "private_dns64_nat_gateway" {
 resource "aws_customer_gateway" "this" {
   for_each = var.customer_gateways
 
-  bgp_asn     = each.value["bgp_asn"]
-  ip_address  = each.value["ip_address"]
-  device_name = lookup(each.value, "device_name", null)
-  type        = "ipsec.1"
+  region = var.region
+
+  bgp_asn          = lookup(each.value, "bgp_asn", null)
+  bgp_asn_extended = lookup(each.value, "bgp_asn_extended", null)
+  ip_address       = each.value["ip_address"]
+  device_name      = lookup(each.value, "device_name", null)
+  type             = "ipsec.1"
 
   tags = merge(
     { Name = "${var.name}-${each.key}" },
     var.tags,
     var.customer_gateway_tags,
   )
+
+  lifecycle {
+    create_before_destroy = true
+  }
 }
 
 ################################################################################
@@ -1131,6 +1313,8 @@ resource "aws_customer_gateway" "this" {
 
 resource "aws_vpn_gateway" "this" {
   count = local.create_vpc && var.enable_vpn_gateway ? 1 : 0
+
+  region = var.region
 
   vpc_id            = local.vpc_id
   amazon_side_asn   = var.amazon_side_asn
@@ -1146,12 +1330,16 @@ resource "aws_vpn_gateway" "this" {
 resource "aws_vpn_gateway_attachment" "this" {
   count = var.vpn_gateway_id != "" ? 1 : 0
 
+  region = var.region
+
   vpc_id         = local.vpc_id
   vpn_gateway_id = var.vpn_gateway_id
 }
 
 resource "aws_vpn_gateway_route_propagation" "public" {
   count = local.create_vpc && var.propagate_public_route_tables_vgw && (var.enable_vpn_gateway || var.vpn_gateway_id != "") ? 1 : 0
+
+  region = var.region
 
   route_table_id = element(aws_route_table.public[*].id, count.index)
   vpn_gateway_id = element(
@@ -1166,6 +1354,8 @@ resource "aws_vpn_gateway_route_propagation" "public" {
 resource "aws_vpn_gateway_route_propagation" "private" {
   count = local.create_vpc && var.propagate_private_route_tables_vgw && (var.enable_vpn_gateway || var.vpn_gateway_id != "") ? local.len_private_subnets : 0
 
+  region = var.region
+
   route_table_id = element(aws_route_table.private[*].id, count.index)
   vpn_gateway_id = element(
     concat(
@@ -1178,6 +1368,8 @@ resource "aws_vpn_gateway_route_propagation" "private" {
 
 resource "aws_vpn_gateway_route_propagation" "intra" {
   count = local.create_vpc && var.propagate_intra_route_tables_vgw && (var.enable_vpn_gateway || var.vpn_gateway_id != "") ? local.len_intra_subnets : 0
+
+  region = var.region
 
   route_table_id = element(aws_route_table.intra[*].id, count.index)
   vpn_gateway_id = element(
@@ -1216,6 +1408,8 @@ resource "aws_vpn_connection_route" "this" {
 resource "aws_default_vpc" "this" {
   count = var.manage_default_vpc ? 1 : 0
 
+  region = var.region
+
   enable_dns_support   = var.default_vpc_enable_dns_support
   enable_dns_hostnames = var.default_vpc_enable_dns_hostnames
 
@@ -1228,6 +1422,8 @@ resource "aws_default_vpc" "this" {
 
 resource "aws_default_security_group" "this" {
   count = local.create_vpc && var.manage_default_security_group ? 1 : 0
+
+  region = var.region
 
   vpc_id = aws_vpc.this[0].id
 
@@ -1274,6 +1470,8 @@ resource "aws_default_security_group" "this" {
 
 resource "aws_default_network_acl" "this" {
   count = local.create_vpc && var.manage_default_network_acl ? 1 : 0
+
+  region = var.region
 
   default_network_acl_id = aws_vpc.this[0].default_network_acl_id
 
@@ -1327,6 +1525,8 @@ resource "aws_default_network_acl" "this" {
 
 resource "aws_default_route_table" "default" {
   count = local.create_vpc && var.manage_default_route_table ? 1 : 0
+
+  region = var.region
 
   default_route_table_id = aws_vpc.this[0].default_route_table_id
   propagating_vgws       = var.default_route_table_propagating_vgws