+++ modules/vpc-endpoints/main.tf
@@ -9,10 +9,14 @@ locals {
 }
 
 data "aws_vpc_endpoint_service" "this" {
-  for_each = local.endpoints
+  # This data source is sort of useless without the following
+  # https://github.com/hashicorp/terraform-provider-aws/issues/42462
+  # It only works in the same region as the provider, regardless of the arguments provided (service, service_name, service_regions, etc.)
+  for_each = { for k, v in local.endpoints : k => v if var.region == null }
 
-  service      = try(each.value.service, null)
-  service_name = try(each.value.service_name, null)
+  service         = try(each.value.service, null)
+  service_name    = try(each.value.service_name, null)
+  service_regions = try([each.value.service_region], null)
 
   filter {
     name   = "service-type"
@@ -23,8 +27,11 @@ data "aws_vpc_endpoint_service" "this" {
 resource "aws_vpc_endpoint" "this" {
   for_each = local.endpoints
 
+  region = var.region
+
   vpc_id            = var.vpc_id
-  service_name      = data.aws_vpc_endpoint_service.this[each.key].service_name
+  service_name      = try(each.value.service_endpoint, data.aws_vpc_endpoint_service.this[each.key].service_name)
+  service_region    = try(each.value.service_region, null)
   vpc_endpoint_type = try(each.value.service_type, "Interface")
   auto_accept       = try(each.value.auto_accept, null)
 
@@ -33,8 +40,32 @@ resource "aws_vpc_endpoint" "this" {
   route_table_ids     = try(each.value.service_type, "Interface") == "Gateway" ? lookup(each.value, "route_table_ids", null) : null
   policy              = try(each.value.policy, null)
   private_dns_enabled = try(each.value.service_type, "Interface") == "Interface" ? try(each.value.private_dns_enabled, null) : null
+  ip_address_type     = try(each.value.ip_address_type, null)
 
-  tags = merge(var.tags, try(each.value.tags, {}))
+  dynamic "dns_options" {
+    for_each = try([each.value.dns_options], [])
+
+    content {
+      dns_record_ip_type                             = try(dns_options.value.dns_record_ip_type, null)
+      private_dns_only_for_inbound_resolver_endpoint = try(dns_options.value.private_dns_only_for_inbound_resolver_endpoint, null)
+    }
+  }
+
+  dynamic "subnet_configuration" {
+    for_each = try(each.value.subnet_configurations, [])
+
+    content {
+      ipv4      = try(subnet_configuration.value.ipv4, null)
+      ipv6      = try(subnet_configuration.value.ipv6, null)
+      subnet_id = try(subnet_configuration.value.subnet_id, null)
+    }
+  }
+
+  tags = merge(
+    var.tags,
+    { "Name" = replace(each.key, ".", "-") },
+    try(each.value.tags, {}),
+  )
 
   timeouts {
     create = try(var.timeouts.create, "10m")
@@ -49,6 +80,8 @@ resource "aws_vpc_endpoint" "this" {
 
 resource "aws_security_group" "this" {
   count = var.create && var.create_security_group ? 1 : 0
+
+  region = var.region
 
   name        = var.security_group_name
   name_prefix = var.security_group_name_prefix
@@ -68,6 +101,8 @@ resource "aws_security_group" "this" {
 
 resource "aws_security_group_rule" "this" {
   for_each = { for k, v in var.security_group_rules : k => v if var.create && var.create_security_group }
+
+  region = var.region
 
   # Required
   security_group_id = aws_security_group.this[0].id